#8: Rename module from ditalib to pydita
Rename the Python package directory from src/ditalib/ to src/pydita/
and update all imports throughout source and test files accordingly.
Update README module table and installation instructions to reflect
the new package name.
