+++ Pipfile
@@ -14,7 +14,7 @@ aenum = "*"
 [dev-packages]
 pytest = "*"
 pylint = "*"
-pydita = {file = ".", editable = true}
+pydita = {path = ".", editable = true}
 
 [requires]
 python_version = "3.10"
