+++ src/pydita/visitor.py
